requirements.txt
argostranslate==1.9.6
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ctranslate2==4.3.1
filelock==3.14.0
fsspec==2024.6.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
libretranslatepy==2.1.1
llvmlite==0.42.0
lxml==5.2.2
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
mysql-connector-python==8.4.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
packaging==24.1
protobuf==5.27.1
pycparser==2.22
pydub==0.25.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
sacremoses==0.0.53
sentencepiece==0.2.0
setuptools==70.1.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
SpeechRecognition==3.10.4
stanza==1.1.1
sympy==1.12.1
tbb==2021.12.0
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
